refactor:共享公开Supabase连接重试器
让完整导入与审阅暂存共同使用可跟踪模块，彻底移除公开CI对本地私有文件的依赖。
